WIP: create restaurants

+++ apps/restify/internal/domain/value_object/hour_minute.go
@@ -6,12 +6,12 @@ import (
 )
 
 type HourMinute struct {
-	Hour   int
-	Minute int
+	Hour   uint8 
+	Minute uint8
 }
 
 // Time format is 24h
-func NewHourMinute(hour, minute int) (*HourMinute, error) {
+func NewHourMinute(hour, minute uint8) (*HourMinute, error) {
 	if hour < 0 || hour > 24 || minute < 0 || minute > 60 {
 		return nil, domain.HourMinuteRangeError
 	}
